backend/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Prashmin Gurung", email = "prashmingrg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlmodel >=0.0.27,<0.0.28",
    "python-dotenv >=1.2.1,<2.0.0",
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"